+++ src/main/java/com/prgrms/ijuju/stock/adv/advancedinvest/dto/response/AdvancedResponseDTO.java
@@ -1,4 +1,4 @@
-package com.prgrms.ijuju.stock.adv.advancedInvest.dto.request;
+package com.prgrms.ijuju.stock.adv.advancedinvest.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,5 +7,5 @@ import lombok.Data;
 @Data
 @Builder
 @AllArgsConstructor
-public class AdvancedRequestDTO {
+public class AdvancedResponseDTO {
 }
