+++ src/main/java/com/prgrms/ijuju/stock/adv/advancedinvest/entity/AdvancedInvest.java
@@ -1,12 +1,9 @@
-package com.prgrms.ijuju.stock.adv.advancedInvest.entity;
+package com.prgrms.ijuju.stock.adv.advancedinvest.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Entity
